+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.1.4] - 2017-01-07
+
 ### Added
 
 - Support for the `arm-unknown-linux-gnueabi` target
@@ -17,6 +19,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - It's no longer necessary to call `cargo generate-lockfile` before using
   `cross` as `cross` will now take care of creating a lockfile when necessary.
+
+- The C environments for the `thumb` targets now include newlib (`libc.a`,
+  `libm.a`, etc.)
+
+### Fixed
+
+- A segfault when `cross` was trying to figure out the name of the user that
+  called it.
 
 ## [v0.1.3] - 2017-01-01
 
@@ -43,7 +53,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release. Supports 12 targets.
 
-[Unreleased]: https://github.com/japaric/cross/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/japaric/cross/compare/v0.1.4...HEAD
+[v0.1.4]: https://github.com/japaric/cross/compare/v0.1.3...v0.1.4
 [v0.1.3]: https://github.com/japaric/cross/compare/v0.1.2...v0.1.3
 [v0.1.2]: https://github.com/japaric/cross/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/japaric/cross/compare/v0.1.0...v0.1.1